Moving app to the lib directory
This commit moves the bulk of Shelltest to the lib directory, but the
constructor stays in the entry point. This will help with testing down
the road.
